refactor: The login page is divided into components
