+++ src/page/login-page/login-form.tsx
@@ -1,0 +1,52 @@
+import { useState, FormEvent } from 'react';
+import { useDispatch } from 'react-redux';
+import { fetchLogin } from '../../store/user/action';
+import { AppDispatchType } from '../../store';
+
+function LoginForm() {
+  const dispatch: AppDispatchType = useDispatch();
+
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleSubmit = (evt: FormEvent) => {
+    evt.preventDefault();
+    dispatch(fetchLogin({ email, password }));
+  };
+
+  return (
+    <form className="login__form form" onSubmit={handleSubmit}>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">E-mail</label>
+        <input
+          className="login__input form__input"
+          type="email"
+          name="email"
+          placeholder="Email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+        />
+      </div>
+
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">Password</label>
+        <input
+          className="login__input form__input"
+          type="password"
+          name="password"
+          placeholder="Password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+        />
+      </div>
+
+      <button className="login__submit form__submit button" type="submit">
+        Sign in
+      </button>
+    </form>
+  );
+}
+
+export default LoginForm;
